refactor(api): split monolithic Api class into grouped sub-APIs (#144)
The Api class in gatox/github/api.py was a 1912-line monolith holding
every GitHub REST/GraphQL call gato-x makes. Split it into a small
orchestrator plus six grouped sub-APIs that share a single HTTP client
+ rate-limit + token state via a new ApiBase infra class:

  RepoApi      (17 methods) — repo metadata, contents, refs, runners,
                              artifacts, run logs at the repo level
  ActionApi    (14 methods) — Actions runners list, workflow runs,
                              run logs, artifacts, environments,
                              workflow_dispatch / repository_dispatch
  CommitApi    (10 methods) — refs, branches, commit metadata,
                              tree mutations, file-last-updated history
  OrgApi       (7 methods)  — org-level reads, runners, app installs
  UserApi      (7 methods)  — current-user / user-by-name endpoints
  AppApi       (7 methods)  — GitHub App installation flow used by
                              the --machine flag

Wire-level behaviour is unchanged: each sub-API holds the same shared
ApiBase (httpx client, headers, rate-limit state, trace2 hooks), so
HTTP traffic / proxy config / SOCKS / etc. all flow through one place
exactly as before.

Migration: every call site in gatox/{enumerate,workflow_graph,attack,
search,cli} updated from `self.api.foo()` to `self.api.<group>.foo()`.
67 call-site changes across 16 files — purely mechanical.

Lifecycle (`__aenter__`/`__aexit__`/`close`) and rate-limit logic
remain on the parent Api class. Sub-APIs are constructed once in
`Api.__init__` and exposed as `self.repo`, `self.org`, etc., so
callers get autocomplete and pyright keeps strong typing through.

Verified:
  pytest -q     -> 556 passed (no change vs baseline)
  pyright       -> 0 errors / 0 warnings / 0 informations
  ruff check    -> all checks passed
  ruff format   -> all formatted

Closes #144.

Follow-up (separate PR): `LocalApiStub` in
gatox/enumerate/local_enumerate.py currently subclasses `Api` to
satisfy pyright after PR #258. Once this lands, LocalApiStub will need
~30 lines to expose stub sub-API attributes (`self.repo = LocalRepoApiStub(...)`
etc.) so the same type-compatibility holds.
